facenet-pytorch/facenet-attendance-system/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyMySQL==1.1.0
cryptography==41.0.7
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
matplotlib==3.8.2
scikit-learn==1.3.2
openpyxl==3.1.2
werkzeug==2.3.7
bcrypt==4.1.1
